chore: rename build action (#245)
We modified the build action in this PR to use matrix: #239 . That caused the Github build status not to be reported correctly; since we are now reporting build (ubuntu-latest) and build(windows-latest) instead of just one build. This PR fixes that so now we are reporting just one build status.