fix(model-library): robustify and refactor (#274)
Background:
Simulator connector needs models to run the simulations on. Each model
is stored on the local disk and every time user uploads a new version,
it gets pulled by the connector and opened in the simulator for parsing
/ verification.
Also, if the user triggers the simulation, the model needs to be
available beforehand. This may happen either before the model has been
processed, after, or even in the middle.

Why need the change:

Currently there are a number of things that arent ideal with the model
file library:
1. When multiple threads access the same model revision, we download a
file again a put it into a temporary folder. This adds more work to the
connector, especially when things start to pile up.
2. Accesing model revision by it's externalId has major flow when user
deletes and reapploads the model with the same externalId. While the
code might handle some if this situations, some edge case might arise
due to the code complexity.
3. Convoluted and hard to debug logic. People reporting files missing,
etc.
4. We have a feature requests and with the current state of the library
it is hard to make such changes.

Changes:

1. Remove the temporary state, if the model revision is being processed
by the connector while other thread tries to access it we simply return
the same singleton task for both threads.
2. Always compare modelRevision.externalId with the remote, if they are
different -> process from scratch
3. Process all the models on the startup. Download a missing file on a
fly when simulation runs happens and the file is not available.
4. Removed the "fail large file" logic, this was needed only for the
files that were put into temporary state.
5. Added extensive test that checks for the correct multi-thread access
to the model.

Based on top of:
https://github.com/cognitedata/dotnet-simulator-utils/pull/279

NOTE: I apologise for numerous changes, but this code is so convoluted
that it's impossible to achieve something by changing a single thing.

---------

Co-authored-by: Luis Pereira <luis.pereira@cognite.com>
Co-authored-by: abdullah-cognite <100700755+abdullah-cognite@users.noreply.github.com>