feat(model-library): support original file extension [POFSP-1111] (#281)
Stacked on top of:
https://github.com/cognitedata/dotnet-simulator-utils/pull/274 (more
changes will follow)

Why:
This is both a logical improvement (paying the tech dept) but also an
feature that unblocks our current inititive (support model external
dependencies).

What:
So far we had been locking file extension to the "first item in the list
of supported file extesions for a given simulator".


This has some major downsides:
1. Actual file extesion is completely ignored
3. Not possible to support multiple file extensions per simulator
(needed soon for the new feature)

Changes:
1. do not call simulators API, rather take the extension from the name
of the file (this is something we validate when user uploads the model)