test: improved parallelism & more convenient test utilities (#283)
The is a prerequisite for another change.
1. We currently cannot concurrently run tests that use LiteDb / disk,
since they step on each other toes.
It's somewhat okay for e2e integration tests, but there is no reason we
shoud not be able to run unit tests (with mocked http client) in
parallel.
This fixes that by adding a postfix to the file / folder being used by
the test.
2. Added convenience utilities AddMockedHttpClientFactory and
AddDefaultConfig to simplify test boilerplate

Also added cleanup step for
[ModelLibraryConcurrencyTest](https://github.com/cognitedata/dotnet-simulator-utils/pull/283/commits/c62f63ce079cea084d3b1efb153f5b31ed60d18c)
(from prev PR)