feat(model-library): support multi-file model downloads and parsing (#282)
This is an unfinished feature and still needs polishing and robustness,
but the PR is already bit enough, so the improvements will come stacked.

There are a few TODOs in the code - those are tickets for this/next
sprint.

**WHY**
We are building a feature would allow support of large model file (dozen
of GBs) for specific simulator connectors with a portial update
capability. The way this is done is that one model revision can
reference multiple smaller files as dependencies. So when one of the
dependencies changed, the connector would not need to download the whole
large model, but just a small part of it.

https://cognitedata.atlassian.net/wiki/spaces/PD/pages/5216337994/Thoughts+on+how+to+support+external+dependencies


**What changed**
This PR instroduces support for the changes in the API
https://api-docs.cognite.com/internal/tag/Simulator-Models/operation/create_simulator_model_revision_simulators_models_revisions_post
One the model revision is read from the API, each file in the list is
then downloaded for each model and the Path is saved to the local state
`DependencyFiles` property (in memory + LiteDB).

There will be more things needed here (e.g. deduplication, if a model
revision needs a file that is already downloaded, batching, etc).

We don't release the library just yet, so this feature is not going to
be used by anyone yet.
<img width="789" height="479" alt="Screenshot 2025-08-06 at 14 22 06"
src="https://github.com/user-attachments/assets/d920e89f-0f6d-4cdf-bad0-50ddf1a29ab7"
/>