feat(modellib): deduplicate file download [POFSP-1137] (#284)
After the recent change for multi-file model support
https://github.com/cognitedata/dotnet-simulator-utils/pull/282 we need
to add some logic for performance optimisation.
The model files (and their dependencies) need to be deduplicated on
download. That means if there is the same file locally already, we don't
have to download it again, and only link it to the existing `Path`.
To achieve that we compare the list of files in the state store to the
files needed for a given model.

Additional change: include `.Downloaded` check for the model deps, in
addition to the main file.