feat(modellib): restore extraction state on startup (#289)
WHY

When we worked on the recent refactor
https://github.com/cognitedata/dotnet-simulator-utils/pull/274 we
noticed that we can do some improvements on how we cache/restore to/from
local state store (LiteDB)
So far the we restored state from local db store into a memory only
after we read all the remote objects from the API. This is unfortunate,
as it won't allow us to figure out what models need to be
downloaded/processed and due to the recent refactor lead to a major flaw
where all the models were re-processed even before the state was
restored.

Changes:
1. Restore the in-memory state from LiteDB before doing any API calls
2. [additional change] Save ModelExternalId into the state store (this
was missing before)
3. [additional change] Remove redundant `SimulatorModelInfo Model`
property (not used), same as `InformationExtracted`. There are no other
references to them currently

Now we also added the test to verify that restoration works the way we
expect with no extra/redundant API calls/processing.