feat(modellib): improve startup load performance (#290)
There is a special case when connector starts up and local state is full
of already downloaded/parsed model revisions.
To figure out if those revisions are up to date with the remote, the
ModelLibrary would go through them and get by id each to compare.

We have quite strict limits on the API (1000 rpm), so when a lot of
revisions are already up to date (noop) this leads to a bottleneck by
being rate limited. This change introduces a 1m MemoryCache that is used
for this specific scenario (when model revision is just fetched).

This has quite a limited lifetime (1 minute) in case there is actual job
to be done (e.g. model revision differs from the local), in that case
it's better to fetch one-by one to not face any outdated entries.
Also list endpoint doesn't give all the fields (e.g. model
dependencies), so in case connector needs it, it won't use the cached
model revision.