feat(modellib): support more than 1000 files per model (#291)
One model can consist of 1000+ files. Currently we download them one by
one to not stress the API and this PR doesn't change the behaviour.

One limitation that we have is that `files.RetrieveAsync` doesn't
support getting metadata of more than 1000 files at once.
This is a safe operation to do concurrently, so I am adding an extension
method to support chunking.

I replaced the method call, so the lines are covered by the previous
tests, though I added one more for a happy path as well.