fix(routine-library): optimise memory consumption [POFSP-1221] (#304)
So far we've been saving all the latest version of routines in memory.
This used to work fine and was needed to read the cron params and
schedule routine executions based on that. This in-memory cache was also
used to run the simulation.

Going forward we might start having larger (few MBs routine revisions)
or having simply much larger amount of routines in total.
So to eliminate any future out-of-memory issues with change the logic to
only store the most important fields (`IncludeAllFields=false` while
listing the latest versions).
Afterwards we store the relevant fields into
`SimulatorRoutineRevisionInfo` abstraction.

In order to run a simulation, we'll make a direct call to the API
instead of fetching the routine from memory.

Tests: I consider this being covered by the current integration test
suite as it's mostly a refactor